etl/requirements.txt
python-dotenv==1.0.0
pydantic==1.10.7
psycopg2-binary==2.9
elasticsearch==8.7.0